Homework 52

+++ src/main/java/com/example/jobsearch/service/impl/ResumeServiceImpl.java
@@ -4,8 +4,12 @@ import com.example.jobsearch.dao.ResumeDao;
 import com.example.jobsearch.dto.ResumeDto;
 import com.example.jobsearch.exception.UserNotFoundException;
 import com.example.jobsearch.model.Resume;
+import com.example.jobsearch.service.CategoryService;
 import com.example.jobsearch.service.ResumeService;
+import com.example.jobsearch.service.UserService;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -15,15 +19,23 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ResumeServiceImpl implements ResumeService {
     private final ResumeDao resumeDao;
+    private final UserService userService;
+    private final CategoryService categoryService;
+
+    @Override
+    public List<ResumeDto> getResumes() {
+        List<Resume> resumes = resumeDao.getResumes();
+        return getResumeDtos(resumes);
+    }
 
     @Override
     public ResumeDto getResumeById(int id) throws UserNotFoundException {
         Resume resume = resumeDao.getResumeById(id).orElseThrow(() -> new UserNotFoundException("Can not find resume with id: " + id));
         return ResumeDto.builder()
                 .id(resume.getId())
-                .userId(resume.getUserId())
+                .user(userService.getUserById(resume.getUserId()))
                 .name(resume.getName())
-                .categoryId(resume.getCategoryId())
+                .category(categoryService.getCategoryById(resume.getCategoryId()))
                 .salary(resume.getSalary())
                 .isActive(resume.getIsActive())
                 .createdDate(resume.getCreatedDate())
@@ -42,13 +54,14 @@ public class ResumeServiceImpl implements ResumeService {
         return getResumeDtos(resumes);
     }
 
+
     private List<ResumeDto> getResumeDtos(List<Resume> resumes) {
         List<ResumeDto> dtos = new ArrayList<>();
         resumes.forEach(e -> dtos.add(ResumeDto.builder()
                 .id(e.getId())
-                .userId(e.getUserId())
+                .user(userService.getUserById(e.getUserId()))
                 .name(e.getName())
-                .categoryId(e.getCategoryId())
+                .category(categoryService.getCategoryById(e.getCategoryId()))
                 .salary(e.getSalary())
                 .isActive(e.getIsActive())
                 .createdDate(e.getCreatedDate())
@@ -57,40 +70,74 @@ public class ResumeServiceImpl implements ResumeService {
         return dtos;
     }
 
+    @SneakyThrows
+    private boolean isEmployee(int userId) {
+        return "Соискатель".equalsIgnoreCase(userService.getUserById(userId).getAccountType());
+    }
+
+    @Override
+    public HttpStatus createResume(ResumeDto resume) {
+        if (isEmployee(resume.getUser().getId())) {
+            resumeDao.createResume(resume);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.BAD_REQUEST;
+    }
+
+    @Override
+    public Boolean isResumeInSystem(int id) {
+        return resumeDao.isResumeInSystem(id);
+    }
+
+    @Override
+    public HttpStatus changeResumeName(int id, String name) {
+        if (isResumeInSystem(id)) {
+            resumeDao.changeResumeName(id, name);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.BAD_REQUEST;
+    }
+
+    @Override
+    public HttpStatus changeResumeCategory(int id, String category) {
+        if (isResumeInSystem(id)) {
+            resumeDao.changeResumeCategory(id, category);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.BAD_REQUEST;
+    }
+
+    @Override
+    public HttpStatus changeResumeSalary(int id, Double salary) {
+        if (isResumeInSystem(id)) {
+            resumeDao.changeResumeSalary(id, salary);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.BAD_REQUEST;
+    }
+
+    @Override
+    public HttpStatus changeResumeActive(int id, Boolean status) {
+        if (isResumeInSystem(id)) {
+            resumeDao.changeResumeActive(id, status);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.BAD_REQUEST;
+    }
+
+    @Override
+    public HttpStatus deleteResumeById(int id) {
+        if (isResumeInSystem(id)) {
+            resumeDao.deleteResumeById(id);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.BAD_REQUEST;
+    }
+
     @Override
     public List<ResumeDto> getActiveResumes(int userId) {
         // TODO реализовать выборку активных резюме
         return null;
-    }
-
-    @Override
-    public void changeResumeName(int id, String name) {
-        // TODO реализовать смену имени
-    }
-
-    @Override
-    public void changeResumeCategory(int id, String category) {
-        // TODO реализовать смену категории
-    }
-
-    @Override
-    public void changeResumeSalary(int id, Double salary) {
-        // TODO реализовать смену зарплаты
-    }
-
-    @Override
-    public void changeResumeActive(int id, Boolean status) {
-        // TODO реализовать смену статуса активности
-    }
-
-    @Override
-    public void createResume(ResumeDto resume) {
-        resumeDao.createResume(resume);
-    }
-
-    @Override
-    public void deleteResumeById(int id) {
-        resumeDao.deleteResumeById(id);
     }
 
 }
